fmt + moved test to its file
