+++ contracts/external/dao-migrator/src/testing/setup.rs
@@ -219,7 +219,7 @@ pub fn init_dao_v1(mut app: App, data: Option<InitDaoDataV1>) -> (App, Addr, Add
     (app, core_addr, proposal_addr, v1_code_ids)
 }
 
-fn execute_migration(
+pub fn execute_migration(
     mut app: App,
     core_addr: Addr,
     proposal_addr: Addr,
@@ -301,17 +301,4 @@ fn execute_migration(
         &cw_proposal_single_v1::msg::ExecuteMsg::Execute { proposal_id: 2 },
         &[],
     )
-}
-
-#[test]
-fn test_migration_v1_v2() {
-    let app = App::default();
-
-    // ----
-    // instantiate a v1 DAO
-    // ----
-    let (app, core_addr, proposal_addr, v1_code_ids) = init_dao_v1(app, None);
-
-    let res = execute_migration(app, core_addr, proposal_addr, v1_code_ids).unwrap();
-    println!("{:?}", res)
 }
